+++ build_daemon/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Report OptionsSkew.
 - Prefix build daemon directory with username.
 - Forward filesystem changes to daemon builder.
+- Pass custom client options to daemon builder.
 
 ## 0.0.1